[RdKafka] Enable serializers to serialize message keys
Currently the `RdKafkaProducer` is serializing just the message, which
I guess us fine for JSON serialization. But when you want to use an
AvroSerializer that integrates with the Confluent Platform, it is
possible to use complex types as keys as well, which in turn means that
keys need to be serialized, too.

This change leverages the mutability of the Message by shifting the
`getKey` method call after the call to the Serializer.